server/py_requirements.txt
flask
gitpython
requests
ipdb
pyyaml
